% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\docType{data}
\name{furmanu_colors}
\alias{furmanu_colors}
\title{Furman University brand colors (hex codes)}
\format{
An object of class \code{character} of length 10.
}
\usage{
furmanu_colors
}
\description{
A named vector of Furman University color hex codes.
}
\keyword{datasets}
