% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\name{furmanu_cols}
\alias{furmanu_cols}
\title{Extract Furman University colors by name}
\usage{
furmanu_cols(...)
}
\arguments{
\item{...}{Character names of Furman colors to extract}
}
\value{
A character vector of hex codes
}
\description{
Extract Furman University colors by name
}
