% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\name{furmanu_pal}
\alias{furmanu_pal}
\title{Interpolated color palette from Furman palette}
\usage{
furmanu_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette to use}

\item{reverse}{Logical. Should the palette be reversed?}

\item{...}{Additional arguments passed to colorRampPalette()}
}
\value{
A function to generate interpolated colors
}
\description{
Interpolated color palette from Furman palette
}
