% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\docType{data}
\name{furmanu_palettes}
\alias{furmanu_palettes}
\title{Furman University color palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
furmanu_palettes
}
\description{
A list of named palettes using furmanu colors
}
\keyword{datasets}
