% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\name{open_furmanu_demo}
\alias{open_furmanu_demo}
\title{Open the Furman ggplot2 theme demo Quarto file}
\usage{
open_furmanu_demo()
}
\description{
This function opens the demo Quarto (.qmd) file included in the themefurmanu package.
It demonstrates various uses of \code{theme_furmanu()} and the \verb{scale_*_furmanu()} functions.
}
\details{
The demo file is located in the installed package directory under inst/demo/demo.qmd.
}
