% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\name{scale_fill_furmanu}
\alias{scale_fill_furmanu}
\title{Fill scale using Furman color palettes}
\usage{
scale_fill_furmanu(
  palette = "main",
  discrete = TRUE,
  reverse = FALSE,
  midpoint = NULL,
  ...
)
}
\arguments{
\item{palette}{Palette name}

\item{discrete}{Is the scale discrete?}

\item{reverse}{Reverse the palette?}

\item{midpoint}{Optional midpoint value for diverging scales}

\item{...}{Other arguments to scale functions}
}
\description{
Fill scale using Furman color palettes
}
