% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furmanu_colors.R
\name{show_furmanu_palettes}
\alias{show_furmanu_palettes}
\title{Display all Furman palettes visually}
\usage{
show_furmanu_palettes()
}
\description{
Shows all named Furman palettes as horizontal bars.
}
